Refactor to cloud package
Separate cloud metadata retrieval into its own package for
readability/maintainability.

Also, removes the 'vendor_data' key from digitalocean data, as it
is quite noisy and breaks shell output.
